stage2/requirements.txt
# Requirements for Stage 2 - Blockchain Engineer

# Core Web3 Libraries
web3>=6.0.0
eth-account>=0.9.0
eth-keys>=0.4.0
eth-typing>=3.0.0
eth-utils>=2.0.0

# Ethereum Node Interaction
eth-keyfile>=0.6.1
rlp>=3.0

# Cryptography & Hashing
cryptography>=40.0.0
pycryptodome>=3.18.0

# Testing & Development
pytest>=7.0.0
pytest-cov>=4.0.0

# Solidity Compilation (optional)
py-solc-x>=1.1.1

# HTTP Client
requests>=2.28.0

# JSON & Data
aiohttp>=3.8.0

# Development Tools
black>=23.0.0
flake8>=5.0.0
pylint>=2.16.0

# Environment Variables
python-dotenv>=1.0.0

# Hardhat Integration (optional, for Node.js projects)
# npm install -g hardhat @nomicfoundation/hardhat-toolbox
